feat: refactor session management to support multiple storage backends
- Extract session implementations into separate classes (KV, DynamoDB, Durable Objects)
- Add DurableObjectSession for Cloudflare Durable Objects storage
- Add DynamoSession for AWS DynamoDB storage
- Add KVSession for Cloudflare KV storage
- Update session middleware to use Durable Objects by default
- Add Durable Objects binding configuration in wrangler.toml
- Export DurableObjectBase from main application entry point
- Update environment interface to include SESSION Durable Object namespace
