+++ src/index.tsx
@@ -62,4 +62,5 @@ app.get(
 );
 app.get(EndpointPath.serviceJwksPath, ServiceJwksController.handle);
 
+export { DurableObjectBase } from './session/DurableObjectSession';
 export default app;
